+++ collections.ts
@@ -1,4 +1,34 @@
 /**
+ * Turns a synchrnous iterable `source` into an async iterable.
+ *
+ * ``` typescript
+ * import { fromIterable } from "./collections.ts";
+ *
+ * function* iterable() { yield 1; yield 2; yield 3; }
+ * const asyncIterable = fromIterable(iterable());
+ * for await (const value of asyncIterable) console.log(value);
+ * ```
+ *
+ * The above example will print the following lines:
+ *
+ * ~~~
+ * 1
+ * 2
+ * 3
+ * ~~~
+ *
+ * @param source The synchronous iterable to take elements from.
+ *               It can be either finite or infinite.
+ * @returns An async iterable that yields the same elements as the `source`
+ *          iterable.
+ */
+export async function* fromIterable<T>(
+  source: Iterable<T>,
+): AsyncIterableIterator<T> {
+  yield* source;
+}
+
+/**
  * Creates an array from an async iterable.
  *
  * ``` typescript