+++ mod.ts
@@ -1,4 +1,4 @@
-export { toArray } from "./collections.ts";
+export { fromIterable, toArray } from "./collections.ts";
 export { reduce } from "./fold.ts";
 export { count, cycle, repeat } from "./infinite.ts";
 export { take, takeWhile } from "./take.ts";